add: capability to define default behaviour to dialogue edges
