+++ python/examples/17-stateful-communication/dialogues/chitchat.py
@@ -1,9 +1,38 @@
-"""Specific dialogue class for the chit-chat dialogue."""
+"""
+Specific dialogue class for the chit-chat dialogue.
+
+The contents of this file are to be shared between the agents that want to
+use this dialogue. This defines the structure of the specific dialogue and
+the messages that are expected to be exchanged.
+"""
 
 from typing import Type
 
 from uagents import Model
+from uagents.context import Context
 from uagents.experimental.dialogues import Dialogue, Edge, Node
+
+
+# define dialogue messages; each transition needs a separate message
+class InitiateChitChatDialogue(Model):
+    pass
+
+
+class AcceptChitChatDialogue(Model):
+    pass
+
+
+class ChitChatDialogueMessage(Model):
+    text: str
+
+
+class ConcludeChitChatDialogue(Model):
+    pass
+
+
+class RejectChitChatDialogue(Model):
+    pass
+
 
 # Node definition for the dialogue states
 default_state = Node(
@@ -64,6 +93,20 @@ end_session = Edge(
     parent=chatting_state,
     child=end_state,
 )
+
+
+# define default behaviour for individual dialogue edges
+async def reject_dialogue(
+    ctx: Context,
+    sender: str,
+    _msg: InitiateChitChatDialogue,
+):
+    """This function is the default behaviour of this specific step in the diaglogue."""
+    ctx.logger.debug("Automatically reject Dialogue request.")
+    await ctx.send(sender, RejectChitChatDialogue())
+
+
+init_session.set_default_behaviour(InitiateChitChatDialogue, reject_dialogue)
 
 
 class ChitChatDialogue(Dialogue):
